requirements.txt
asgiref==3.5.0
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-js-asset==1.2.2
django-storages==1.12.3
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
gunicorn==20.1.0
idna==3.3
install==1.3.5
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
oauthlib==3.2.0
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
whitenoise==5.3.0
